pyproject.toml
[project]
name = "voiceapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
# TODO: After js frontend pyzmq in dependencies for all
dependencies = [
    "numpy==1.26.4",
    "pydub>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit==4.1.0",
    "ruff==0.11.0",
]

fastapi = [
    "fastapi==0.115.2",
    "uvicorn==0.30.6",
    "pyzmq==26.3.0",
    "websockets==15.0.1",
    "sqlalchemy>=2.0.40",
    "psycopg2>=2.9.10",
]

sherpa = [
    "pyzmq==26.3.0",
    "onnxruntime-gpu==1.19.2",
    "soundfile==0.12.1",
    "scipy==1.13.1",
]

frontend = [
    "gradio==5.20.0",
    "streamlit==1.43.2",
    "pyaudio==0.2.14",
    "websockets==15.0.1",
]

agent = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-google-genai>=2.1.3",
    "langchain-tavily>=0.1.5",
    "langgraph>=0.3.30",
    "python-dotenv>=1.1.0",
    "pyzmq==26.3.0",
]
